+++ src/components/reactions/ReactionList.js
@@ -1,0 +1,3 @@
+// Export a reactionList funciton that accepts post as props
+
+    // Return a list of reactions for the incoming post with buttons that add/update/remove reactions
