+++ src/components/reactions/ReactionManager.js
@@ -1,0 +1,1 @@
+// TODO Create createReaction
